requirements.txt
google-adk>=0.5.0
uvicorn[standard]>=0.30.0
fastapi>=0.115.0
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
